LayoutTests/fast/dom/xmlhttprequest-constructor-in-detached-document-expected.txt

CONSOLE MESSAGE: line 12: ReferenceError: XMLHttpRequest constructor associated document is unavailable
Text for bug 25290: Crash when constructing XMLHttpRequest in a detached document.

PASS
